Tests improvements: - moved modules to common place - use ::class notation - use aliases+factories instead of invokables
