--- test/Filter/module/Db/src/Db/Module.php
@@ -1,23 +1,0 @@
-<?php
-namespace Db;
-
-use ZF\Apigility\Provider\ApigilityProviderInterface;
-
-class Module implements ApigilityProviderInterface
-{
-    public function getConfig()
-    {
-        return include __DIR__ . '/../../config/module.config.php';
-    }
-
-    public function getAutoloaderConfig()
-    {
-        return [
-            'Zend\Loader\StandardAutoloader' => [
-                'namespaces' => [
-                    __NAMESPACE__ => __DIR__,
-                ],
-            ],
-        ];
-    }
-}
