+++ test/OrderBy/local.php
@@ -4,13 +4,17 @@
  * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
+use Doctrine\DBAL\Driver\PDOSqlite\Driver;
+use Zend\ServiceManager\Factory\InvokableFactory;
+use ZF\Doctrine\QueryBuilder\OrderBy;
+
 return [
     'doctrine' => [
         'connection' => [
             'orm_default' => [
                 'configuration' => 'orm_default',
                 'eventmanager'  => 'orm_default',
-                'driverClass'   => 'Doctrine\DBAL\Driver\PDOSqlite\Driver',
+                'driverClass'   => Driver::class,
                 'params' => [
                     'memory' => true,
                 ],
@@ -25,13 +29,19 @@ return [
         ],
     ],
     'zf-doctrine-querybuilder-orderby-orm' => [
-        'invokables' => [
-            'field' => 'ZF\Doctrine\QueryBuilder\OrderBy\ORM\Field',
+        'aliases' => [
+            'field' => OrderBy\ORM\Field::class,
+        ],
+        'factories' => [
+            OrderBy\ORM\Field::class => InvokableFactory::class,
         ],
     ],
     'zf-doctrine-querybuilder-orderby-odm' => [
-        'invokables' => [
-            'field' => 'ZF\Doctrine\QueryBuilder\OrderBy\ODM\Field',
+        'aliases' => [
+            'field' => OrderBy\ODM\Field::class,
+        ],
+        'factories' => [
+            OrderBy\ODM\Field::class => InvokableFactory::class,
         ],
     ],
 ];
