--- test/OrderBy/module/Db/Module.php
@@ -1,2 +1,0 @@
-<?php
-require __DIR__ . '/src/Db/Module.php';
