--- test/OrderBy/module/Db/config/xml/Db.Entity.Album.dcm.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xsi="http://www.w3.org/2001/XMLSchema-instance" schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="Db\Entity\Album">
-    <id name="id" type="integer">
-      <generator strategy="AUTO"/>
-    </id>
-    <field name="name" type="string" nullable="true"/>
-    <field name="createdAt" type="datetime" nullable="true"/>
-    <many-to-one field="artist" target-entity="Db\Entity\Artist" inversed-by="album">
-      <join-columns>
-        <join-column name="artist_id" referenced-column-name="id"/>
-      </join-columns>
-    </many-to-one>
-  </entity>
-</doctrine-mapping>
