--- test/OrderBy/module/Db/config/xml/Db.Entity.Artist.dcm.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xsi="http://www.w3.org/2001/XMLSchema-instance" schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="Db\Entity\Artist">
-    <id name="id" type="integer">
-      <generator strategy="AUTO"/>
-    </id>
-    <field name="name" type="string" nullable="true"/>
-    <field name="createdAt" type="datetime" nullable="true"/>
-    <one-to-many field="album" target-entity="Db\Entity\Album" mapped-by="artist"/>
-  </entity>
-</doctrine-mapping>
