--- test/OrderBy/module/DbMongo/Module.php
@@ -1,2 +1,0 @@
-<?php
-require __DIR__ . '/src/DbMongo/Module.php';
