--- test/OrderBy/module/DbMongo/config/yml/DbMongo.Document.Meta.dcm.yml
@@ -1,11 +1,0 @@
-DbMongo\Document\Meta:
-  db: zf_doctrine_querybuilder_filter_test
-  collection: meta
-  repositoryClass: DbMongo\Repository\MetaRepository
-  fields:
-    id:
-      id: true
-    name:
-      type: string
-    createdAt:
-      type: date
