--- test/OrderBy/module/DbMongo/src/DbMongo/Repository/MetaRepository.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DbMongo\Repository;
-
-use Doctrine\ODM\MongoDB\DocumentRepository;
-
-class MetaRepository extends DocumentRepository
-{
-}
