--- test/OrderBy/module/General/Module.php
@@ -1,43 +1,0 @@
-<?php
-namespace General;
-
-use Zend\EventManager\EventInterface;
-use Zend\ModuleManager\Feature\BootstrapListenerInterface;
-use ZF\Apigility\Provider\ApigilityProviderInterface;
-
-class Module implements ApigilityProviderInterface, BootstrapListenerInterface
-{
-    public function getConfig()
-    {
-        return include __DIR__ . '/config/module.config.php';
-    }
-
-    public function getAutoloaderConfig()
-    {
-        return [
-            'Zend\Loader\StandardAutoloader' => [
-                'namespaces' => [
-                    __NAMESPACE__ => __DIR__,
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Add the event catcher
-     *
-     * @param EventInterface $e
-     *
-     * @return array
-     */
-    public function onBootstrap(EventInterface $e)
-    {
-        $application = $e->getApplication();
-        $serviceManager = $application->getServiceManager();
-        $eventManager = $application->getEventManager();
-        $sharedEventManager = $eventManager->getSharedManager();
-
-        $eventCatcher = $serviceManager->get('General\Listener\EventCatcher');
-        $sharedEventManager->attachAggregate($eventCatcher);
-    }
-}
