--- test/OrderBy/module/General/config/module.config.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Db;
-
-return [
-    'service_manager' => [
-        'invokables' => [
-            'General\Listener\EventCatcher' => 'General\Listener\EventCatcher',
-        ],
-    ],
-];
