+++ test/assets/module/Db/src/Db/Entity/Album.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace Db\Entity;
+
+class Album
+{
+    protected $id;
+
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    protected $name;
+
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    public function setName($value)
+    {
+        $this->name = $value;
+
+        return $this;
+    }
+
+    protected $createdAt;
+
+    public function getCreatedAt()
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(\DateTime $value)
+    {
+        $this->createdAt = $value;
+
+        return $this;
+    }
+
+    protected $artist;
+
+    public function getArtist()
+    {
+        return $this->artist;
+    }
+
+    public function setArtist($value)
+    {
+        $this->artist = $value;
+
+        return $this;
+    }
+}
