+++ test/assets/module/Db/view/error/index.phtml
@@ -1,0 +1,59 @@
+<h1>An error occurred</h1>
+<h2><?= $this->message ?></h2>
+
+<?php if (! empty($this->display_exceptions)) : ?>
+
+    <?php if (isset($this->exception)
+        && ($this->exception instanceof \Throwable || $this->exception instanceof \Exception)) : ?>
+        <hr/>
+        <h2>Additional information:</h2>
+        <h3><?= get_class($this->exception) ?></h3>
+        <dl>
+            <dt>File:</dt>
+            <dd>
+                <pre class="prettyprint linenums"><?= $this->exception->getFile() ?>:<?= $this->exception->getLine() ?></pre>
+            </dd>
+            <dt>Message:</dt>
+            <dd>
+                <pre class="prettyprint linenums"><?= $this->escapeHtml($this->exception->getMessage()) ?></pre>
+            </dd>
+            <dt>Stack trace:</dt>
+            <dd>
+                <pre class="prettyprint linenums"><?= $this->exception->getTraceAsString() ?></pre>
+            </dd>
+        </dl>
+        <?php $e = $this->exception->getPrevious() ?>
+        <?php if ($e) : ?>
+            <hr/>
+            <h2>Previous exceptions:</h2>
+            <ul class="unstyled">
+                <?php while ($e) : ?>
+                    <li>
+                        <h3><?= get_class($e); ?></h3>
+                        <dl>
+                            <dt>File:</dt>
+                            <dd>
+                                <pre class="prettyprint linenums"><?= $e->getFile() ?>:<?= $e->getLine() ?></pre>
+                            </dd>
+                            <dt>Message:</dt>
+                            <dd>
+                                <pre class="prettyprint linenums"><?= $this->escapeHtml($e->getMessage()) ?></pre>
+                            </dd>
+                            <dt>Stack trace:</dt>
+                            <dd>
+                                <pre class="prettyprint linenums"><?= $e->getTraceAsString() ?></pre>
+                            </dd>
+                        </dl>
+                    </li>
+                    <?php $e = $e->getPrevious() ?>
+                <?php endwhile ?>
+            </ul>
+        <?php endif ?>
+
+    <?php else : ?>
+
+        <h3>No Exception available</h3>
+
+    <?php endif ?>
+
+<?php endif ?>
